Implement proper first-party artifact mapping
