+++ src/python/pants/backend/helm/resolve/artifacts_test.py
@@ -8,6 +8,7 @@ from textwrap import dedent
 import pytest
 
 from pants.backend.helm.resolve.artifacts import (
+    FirstPartyArtifactMapping,
     HelmArtifact,
     HelmArtifactClassicRepositoryLocation,
     HelmArtifactRegistryLocation,
@@ -15,24 +16,97 @@ from pants.backend.helm.resolve.artifacts import (
     ThirdPartyArtifactMapping,
 )
 from pants.backend.helm.resolve.artifacts import rules as artifacts_rules
-from pants.backend.helm.target_types import AllHelmArtifactTargets, HelmArtifactTarget
+from pants.backend.helm.target_types import (
+    AllHelmArtifactTargets,
+    HelmArtifactTarget,
+    HelmChartTarget,
+)
 from pants.backend.helm.target_types import rules as target_types_rules
+from pants.backend.helm.testutil import (
+    HELM_TEMPLATE_HELPERS_FILE,
+    HELM_VALUES_FILE,
+    K8S_SERVICE_FILE,
+)
 from pants.engine.addresses import Address
+from pants.engine.internals.scheduler import ExecutionError
 from pants.engine.rules import QueryRule
 from pants.testutil.rule_runner import RuleRunner
+from pants.util.strutil import bullet_list
 
 
 @pytest.fixture
 def rule_runner() -> RuleRunner:
     return RuleRunner(
-        target_types=[HelmArtifactTarget],
+        target_types=[HelmArtifactTarget, HelmChartTarget],
         rules=[
             *artifacts_rules(),
             *target_types_rules(),
             QueryRule(AllHelmArtifactTargets, ()),
+            QueryRule(FirstPartyArtifactMapping, ()),
             QueryRule(ThirdPartyArtifactMapping, ()),
         ],
     )
+
+
+def test_build_first_party_mapping(rule_runner: RuleRunner) -> None:
+    rule_runner.write_files(
+        {
+            "src/BUILD": "helm_chart(name='foo')",
+            "src/Chart.yaml": dedent(
+                """\
+                apiVersion: v2
+                name: chart-name
+                version: 0.1.0
+                """
+            ),
+            "src/values.yaml": HELM_VALUES_FILE,
+            "src/templates/_helpers.tpl": HELM_TEMPLATE_HELPERS_FILE,
+            "src/templates/service.yaml": K8S_SERVICE_FILE,
+        }
+    )
+
+    tgt = rule_runner.get_target(Address("src", target_name="foo"))
+    mapping = rule_runner.request(FirstPartyArtifactMapping, [])
+    assert mapping["chart-name"] == tgt.address
+
+
+def test_duplicate_first_party_artifact_mappings(rule_runner: RuleRunner) -> None:
+    rule_runner.write_files(
+        {
+            "src/foo/BUILD": "helm_chart()",
+            "src/foo/Chart.yaml": dedent(
+                """\
+                apiVersion: v2
+                name: chart-name
+                version: 0.1.0
+                """
+            ),
+            "src/foo/values.yaml": HELM_VALUES_FILE,
+            "src/foo/templates/_helpers.tpl": HELM_TEMPLATE_HELPERS_FILE,
+            "src/foo/templates/service.yaml": K8S_SERVICE_FILE,
+            "src/bar/BUILD": "helm_chart()",
+            "src/bar/Chart.yaml": dedent(
+                """\
+                apiVersion: v2
+                name: chart-name
+                version: 0.1.0
+                """
+            ),
+            "src/bar/values.yaml": HELM_VALUES_FILE,
+            "src/bar/templates/_helpers.tpl": HELM_TEMPLATE_HELPERS_FILE,
+            "src/bar/templates/service.yaml": K8S_SERVICE_FILE,
+        }
+    )
+
+    expected_err_msg = (
+        "Found more than one `helm_chart` target using the same chart name:\n\n"
+        f"{bullet_list(['src/bar:bar -> chart-name', 'src/foo:foo -> chart-name'])}"
+    )
+
+    with pytest.raises(ExecutionError) as err_info:
+        rule_runner.request(FirstPartyArtifactMapping, [])
+
+    assert expected_err_msg in err_info.value.args[0]
 
 
 def test_build_third_party_mapping(rule_runner: RuleRunner) -> None:
