+++ src/python/pants/backend/helm/util_rules/chart_test.py
@@ -6,27 +6,22 @@ from __future__ import annotations
 from textwrap import dedent
 
 import pytest
-import yaml
 
 from pants.backend.helm.dependency_inference.chart import rules as chart_infer_rules
-from pants.backend.helm.subsystems import helm
 from pants.backend.helm.target_types import HelmArtifactTarget, HelmChartTarget
 from pants.backend.helm.target_types import rules as target_types_rules
 from pants.backend.helm.testutil import (
-    HELM_CHART_FILE_V1_FULL,
-    HELM_CHART_FILE_V2_FULL,
     HELM_TEMPLATE_HELPERS_FILE,
     HELM_VALUES_FILE,
     K8S_SERVICE_FILE,
     gen_chart_file,
 )
 from pants.backend.helm.util_rules import chart, sources, tool
-from pants.backend.helm.util_rules.chart import (
+from pants.backend.helm.util_rules.chart import HelmChart, HelmChartRequest
+from pants.backend.helm.util_rules.chart_metadata import (
     ChartType,
-    HelmChart,
     HelmChartDependency,
     HelmChartMetadata,
-    HelmChartRequest,
 )
 from pants.build_graph.address import Address
 from pants.core.util_rules import config_files, external_tool, stripped_source_files
@@ -44,7 +39,6 @@ def rule_runner() -> RuleRunner:
             *external_tool.rules(),
             *chart.rules(),
             *chart_infer_rules(),
-            *helm.rules(),
             *sources.rules(),
             *tool.rules(),
             *process.rules(),
@@ -56,25 +50,22 @@ def rule_runner() -> RuleRunner:
 
 
 _TEST_CHART_COLLECT_SOURCES_PARAMS = [
-    ("foo", "0.1.0", ChartType.APPLICATION, "https://www.example.com/icon.png", False),
-    ("bar", "0.2.0", ChartType.LIBRARY, None, True),
+    ("foo", "0.1.0", ChartType.APPLICATION, "https://www.example.com/icon.png"),
+    ("bar", "0.2.0", ChartType.LIBRARY, None),
 ]
 
 
-@pytest.mark.parametrize(
-    "name, version, type, icon, lint_strict", _TEST_CHART_COLLECT_SOURCES_PARAMS
-)
+@pytest.mark.parametrize("name, version, type, icon", _TEST_CHART_COLLECT_SOURCES_PARAMS)
 def test_collects_single_chart_sources(
     rule_runner: RuleRunner,
     name: str,
     version: str,
     type: ChartType,
     icon: str | None,
-    lint_strict: bool,
 ) -> None:
     rule_runner.write_files(
         {
-            "BUILD": f"helm_chart(name='{name}', lint_strict={lint_strict})",
+            "BUILD": f"helm_chart(name='{name}')",
             "Chart.yaml": gen_chart_file(name, version=version, type=type, icon=icon),
             "values.yaml": HELM_VALUES_FILE,
             "templates/_helpers.tpl": HELM_TEMPLATE_HELPERS_FILE,
@@ -201,34 +192,10 @@ def test_gathers_all_subchart_sources_inferring_dependencies(rule_runner: RuleRu
     )
 
     target = rule_runner.get_target(Address("src/chart2", target_name="chart2"))
-    helm_chart = rule_runner.request(
-        HelmChart, [HelmChartRequest.from_target(target, generate_chart_lockfile=True)]
-    )
+    helm_chart = rule_runner.request(HelmChart, [HelmChartRequest.from_target(target)])
 
     assert helm_chart.metadata == expected_metadata
-    assert "chart2/Chart.lock" in helm_chart.snapshot.files
     assert "chart2/charts/chart1" in helm_chart.snapshot.dirs
     assert "chart2/charts/chart1/templates/service.yaml" in helm_chart.snapshot.files
     assert "chart2/charts/cert-manager" in helm_chart.snapshot.dirs
     assert "chart2/charts/cert-manager/Chart.yaml" in helm_chart.snapshot.files
-
-
-_TEST_METADATA_PARSER_PARAMS = [
-    (HELM_CHART_FILE_V1_FULL),
-    (HELM_CHART_FILE_V2_FULL),
-]
-
-
-@pytest.mark.parametrize("chart_file", _TEST_METADATA_PARSER_PARAMS)
-def test_metadata_parser_syntax(chart_file: str) -> None:
-    chart_dict = yaml.safe_load(chart_file)
-    metadata = HelmChartMetadata.from_bytes(chart_file.encode())
-
-    rendered_chart_file = metadata.to_yaml()
-    rendered_chart_dict = yaml.safe_load(rendered_chart_file)
-
-    # Amend the original chart dictionary so the can be safely compared
-    if metadata.api_version == "v1":
-        chart_dict["apiVersion"] = "v1"
-
-    assert chart_dict == rendered_chart_dict
